+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix '1000 Required Fields error' in rare cases when creating a shipping update request
+
 ### Changed
 - Only retrieve issuers for a gateway when it has been activated