+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Fix '1000 Required Fields error' in rare cases when creating a shipping update request
-
+- Make sure checkout loads if API is unreachable
+- 
 ### Changed
 - Only retrieve issuers for a gateway when it has been activated