+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added support for optional customer city argument
+
 ### Fixed
-- Fix '1000 Required Fields error' in rare cases when creating a shipping update request
+- Fixed '1000 Required Fields error' in rare cases when creating a shipping update request
 - Make sure checkout loads if API is unreachable
 - 
 ### Changed